splitting codemap into mouse events and rendering
