+++ visualization/app/codeCharta/ui/codeMap/codeMap.ts
@@ -3,14 +3,20 @@ import angular from "angular";
 import "./threeViewer/threeViewer";
 import "../../core/core.module";
 
-import {CodeMapService} from "./codeMapService";
 import {codeMapComponent} from "./codeMapComponent";
+import {CodeMapMouseEventService} from "./codeMap.mouseEvent.service";
+import {CodeMapRenderService} from "./codeMap.render.service";
 
 angular.module("app.codeCharta.ui.codeMap",["app.codeCharta.ui.codeMap.threeViewer", "app.codeCharta.core"])
-
-.component(codeMapComponent.selector, codeMapComponent)
-
+.component(
+    codeMapComponent.selector,
+    codeMapComponent
+)
 .service(
-    CodeMapService.SELECTOR,
-    CodeMapService
+    CodeMapRenderService.SELECTOR,
+    CodeMapRenderService
+)
+.service(
+    CodeMapMouseEventService.SELECTOR,
+    CodeMapMouseEventService
 );
