+++ visualization/app/codeCharta/ui/codeMap/codeMapComponent.ts
@@ -2,222 +2,26 @@ import {codeMapBuilding} from "./rendering/codeMapBuilding"
 import {ThreeCameraService} from "./threeViewer/threeCameraService";
 import {IAngularEvent, IRootScopeService} from "angular";
 import {ThreeViewerService} from "./threeViewer/threeViewerService";
-import {CodeMapService} from "./codeMapService";
 import {MapTreeViewHoverEventSubscriber, MapTreeViewLevelController} from "../mapTreeView/mapTreeViewLevelComponent";
 import {CodeMapNode} from "../../core/data/model/CodeMap";
+import {CodeMapMouseEventService} from "./codeMap.MouseEvent.service";
+import {CodeMapRenderService} from "./codeMap.render.service";
 
-interface Coordinates {
-    x: number,
-    y: number
-}
-
-export interface CodeMapBuildingTransition {
-    from: codeMapBuilding,
-    to: codeMapBuilding
-}
-
-export interface CodeMapControllerSubscriber {
-    onBuildingHovered(data: CodeMapBuildingTransition, event: IAngularEvent),
-    onBuildingSelected(data: CodeMapBuildingTransition, event: IAngularEvent)
-}
-
-/**
- * Controls the codeMapDirective
- */
-export class CodeMapController implements MapTreeViewHoverEventSubscriber{
-
-    private hovered;
-    private selected;
-    private mouse: Coordinates;
-    private dragOrClickFlag = 0;
+export class CodeMapController {
 
     /* @ngInject */
     constructor(
-        private $rootScope,
-        private threeCameraService: ThreeCameraService,
-        private threeRendererService,
-        private threeSceneService,
-        private threeUpdateCycleService,
         private threeViewerService: ThreeViewerService,
         private $element: Element,
-        private codeMapService: CodeMapService, //we need to call this service somewhere.
+        private codeMapMouseEventService: CodeMapMouseEventService,
+        private codeMapRenderService: CodeMapRenderService, //we need to call this service somewhere.
     ){
-
-        /**
-         * hovered mesh
-         * @type {object}
-         */
-        this.hovered = null;
-
-        /**
-         * selected mesh
-         * @type {object}
-         */
-        this.selected = null;
-
-        /**
-         * current mouse position
-         * @type {{x: number, y: number}}
-         */
-        this.mouse = {x: 0, y: 0};
-
-        threeUpdateCycleService.register(this.update.bind(this));
-        MapTreeViewLevelController.subscribeToHoverEvents($rootScope, this);
     }
 
     $postLink() {
         this.threeViewerService.init(this.$element[0]);
         this.threeViewerService.animate();
-        this.threeRendererService.renderer.domElement.addEventListener("mousemove", this.onDocumentMouseMove.bind(this), false);
-        this.threeRendererService.renderer.domElement.addEventListener("mouseup", this.onDocumentMouseUp.bind(this), false);
-        this.threeRendererService.renderer.domElement.addEventListener("mousedown", this.onDocumentMouseDown.bind(this), false);
-    }
-
-    onShouldHoverNode(node: CodeMapNode) {
-        let buildings: codeMapBuilding[] = this.codeMapService.mapMesh.getMeshDescription().buildings;
-        buildings.forEach((building)=>{
-            if(building.node.path === node.path){
-                this.onBuildingHovered(this.hovered, building);
-            }
-        });
-    }
-
-    onShouldUnhoverNode(node: CodeMapNode) {
-        this.onBuildingHovered(this.hovered, null);
-    }
-
-    static subscribe($rootScope: IRootScopeService, subscriber: CodeMapControllerSubscriber) {
-
-        $rootScope.$on("building-hovered", (e, data:CodeMapBuildingTransition) => {
-            subscriber.onBuildingHovered(data, e);
-        });
-
-        $rootScope.$on("building-selected", (e, data:CodeMapBuildingTransition) => {
-            subscriber.onBuildingSelected(data, e);
-        });
-
-    }
-
-    /**
-     * Update method which is bound to the {@link UpdateCycleService}
-     */
-    update() {
-        this.threeCameraService.camera.updateMatrixWorld(false);
-
-        if (this.threeSceneService.getMapMesh() != null)
-        {
-            let intersectionResult = this.threeSceneService.getMapMesh().checkMouseRayMeshIntersection(this.mouse, this.threeCameraService.camera);
-
-            let from = this.hovered;
-            let to = null;
-
-            if (intersectionResult.intersectionFound == true)
-            {
-                to = intersectionResult.building;
-            }
-            else
-            {
-                to = null;
-            }
-
-            if (from != to)
-            {
-                this.onBuildingHovered(from, to);
-                this.hovered = to;
-            }
-        }
-    }
-
-    /**
-     * updates {CodeMapController.mouse} on mouse movement
-     * @param {MouseEvent} event
-     */
-    onDocumentMouseMove (event) {
-        this.mouse.x = ( event.clientX / this.threeRendererService.renderer.domElement.width ) * 2 - 1;
-        this.mouse.y = -( event.clientY / this.threeRendererService.renderer.domElement.height ) * 2 + 1;
-        this.dragOrClickFlag = 1;
-    }
-
-    /**
-     * updates {CodeMapController} on mouse up
-     */
-    onDocumentMouseUp() {
-
-        if(this.dragOrClickFlag === 0) {
-
-            var from = this.selected;
-
-            if (this.hovered) {
-                this.selected = this.hovered;
-                this.onBuildingSelected(from, this.selected);
-            }
-
-            if(!this.hovered && this.selected) {
-                this.selected = null;
-                this.onBuildingSelected(from, null);
-            }
-
-        } else if (this.dragOrClickFlag === 1){
-            //drag
-        }
-
-    }
-
-    /**
-     * updates {CodeMapController} on mouse down
-     */
-    onDocumentMouseDown() {
-        this.dragOrClickFlag = 0;
-    }
-
-    /**
-     * Called when a building is hovered.
-     * @param from old building
-     * @param to new building
-     * @emits {building-hovered} on hover
-     */
-    onBuildingHovered(from: codeMapBuilding, to: codeMapBuilding){
-        /*
-         if the hovered node does not have useful data, then we should look at its parent. If the parent has useful data
-         then this parent is a delta node which is made of two seperate, data-free nodes. This quick fix helps us to
-         handle delta objects, until there is a method for mergng their meshes and materials correctly.
-         See codeMapService.js
-         */
-        if(to && !to.node){
-            if(to.parent && to.parent.node){
-                to.node = to.parent.node;
-            }
-        }
-
-        this.$rootScope.$broadcast("building-hovered", {to: to, from: from});
-
-        if (to !== null)
-        {
-            this.threeSceneService.getMapMesh().setHighlighted([to]);
-        }
-        else
-        {
-            this.threeSceneService.getMapMesh().clearHighlight();
-        }
-    }
-
-    /**
-     * called when a building is selected.
-     * @param from previously selected building
-     * @param to currently selected building
-     * @emits {building-selected} when building is selected
-     */
-    onBuildingSelected(from: codeMapBuilding, to: codeMapBuilding){
-        this.$rootScope.$broadcast("building-selected", {to: to, from:from});
-
-        if (to !== null)
-        {
-            this.threeSceneService.getMapMesh().setSelected([to]);
-        }
-        else
-        {
-            this.threeSceneService.getMapMesh().clearSelected();
-        }
+        this.codeMapMouseEventService.start();
     }
 
 };
